Add support for PlatformColor
