+++ src/ios/LottieReactNative/LRNAnimationViewManagerObjC.m
@@ -1,4 +1,5 @@
 #import <React/RCTViewManager.h>
+#import "RCTConvert+Lottie.h"
 
 @interface RCT_EXTERN_MODULE(LottieAnimationView, RCTViewManager)
 
@@ -9,7 +10,7 @@ RCT_EXPORT_VIEW_PROPERTY(progress, CGFloat);
 RCT_EXPORT_VIEW_PROPERTY(loop, BOOL);
 RCT_EXPORT_VIEW_PROPERTY(speed, CGFloat);
 RCT_EXPORT_VIEW_PROPERTY(onAnimationFinish, RCTBubblingEventBlock);
-RCT_EXPORT_VIEW_PROPERTY(colorFilters, NSArray);
+RCT_EXPORT_VIEW_PROPERTY(colorFilters, LRNColorFilters);
 
 RCT_EXTERN_METHOD(play:(nonnull NSNumber *)reactTag fromFrame:(nonnull NSNumber *) startFrame toFrame:(nonnull NSNumber *) endFrame);
