+++ src/ios/LottieReactNative/RCTConvert+Lottie.h
@@ -1,0 +1,7 @@
+#import <React/RCTConvert.h>
+
+@interface RCTConvert (Lottie)
+
++ (NSArray<NSDictionary *> *)LRNColorFilters:(id)json;
+
+@end
